refactor: move the non-extproc packages out of internal/extproc
