+++ internal/extproc/rerank_processor_test.go
@@ -20,13 +20,13 @@ import (
 	"github.com/stretchr/testify/require"
 
 	cohere "github.com/envoyproxy/ai-gateway/internal/apischema/cohere"
-	"github.com/envoyproxy/ai-gateway/internal/extproc/translator"
 	"github.com/envoyproxy/ai-gateway/internal/filterapi"
 	"github.com/envoyproxy/ai-gateway/internal/headermutator"
 	"github.com/envoyproxy/ai-gateway/internal/internalapi"
 	"github.com/envoyproxy/ai-gateway/internal/llmcostcel"
 	"github.com/envoyproxy/ai-gateway/internal/metrics"
 	tracing "github.com/envoyproxy/ai-gateway/internal/tracing/api"
+	"github.com/envoyproxy/ai-gateway/internal/translator"
 )
 
 func TestRerank_Schema(t *testing.T) {
