+++ internal/translator/gemini_helper.go
