+++ internal/translator/gemini_helper_test.go
