+++ internal/translator/imagegeneration_openai_openai.go
