+++ internal/translator/imagegeneration_openai_openai_test.go
