+++ internal/translator/jsonschema_helper.go
