+++ internal/translator/jsonschema_helper_test.go
