+++ internal/translator/openai_azureopenai_embeddings.go
