+++ internal/translator/openai_completions.go
