+++ internal/translator/openai_completions_test.go
