+++ internal/translator/openai_embeddings.go
