+++ internal/translator/openai_gcpanthropic.go
