+++ internal/translator/openai_gcpanthropic_stream.go
