+++ internal/translator/openai_gcpvertexai.go
