+++ internal/translator/openai_gcpvertexai_test.go
