+++ internal/translator/openai_openai.go
