+++ internal/translator/util.go
